requirements.txt
# Core dependencies
selenium>=4.18.1
python-dotenv>=1.0.1
webdriver-manager>=4.0.1
streamlit>=1.32.2

# Database and Analysis
chromadb
numpy
plotly>=5.19.0
pandas>=2.1.4

# LLM and API
langchain>=0.1.12
langchain-community>=0.1.0
openai>=1.14.1
requests>=2.31.0

# Utilities
logging>=0.4.9.6
beautifulsoup4>=4.12.3
fake-useragent>=1.4.0

# Document Processing
fpdf
python-docx>=1.0.1
PyPDF2>=3.0.1
pdfkit>=1.0.0
tika>=2.6.0

# Configuration
pydantic>=2.6.3
pydantic-settings>=2.2.1